Add more react elements
This commit is to add more of the elements needed to start the react
app. I think a lot of it is probably pretty unnecessary, so I will
definitely be deleting it later. I am switching directions to mainly
focusing on taking this from the react forward approach rather than
the php forward approach I was doing initially.